Add "script_helper.py" to project

+++ Convert to MP3.py
@@ -1,16 +1,14 @@
 #!/usr/bin/python3
-__author__ = 'diogo'
+__author__ = "Diogo Machado"
+import script_helper
 import subprocess
 import time
 import os
 import os.path
-import dbus
-session_bus = dbus.SessionBus()
-proxy = session_bus.get_object('org.freedesktop.Notifications','/org/freedesktop/Notifications')
-notification = dbus.Interface(proxy, 'org.freedesktop.Notifications')
 
-file_paths = os.environ['NAUTILUS_SCRIPT_SELECTED_FILE_PATHS']
-file_paths = file_paths.splitlines()
+notification = script_helper.get_notification_interface()
+file_paths = script_helper.get_selected_files()
+
 files_not_converted = []
 for file in file_paths:
     file_name = os.path.basename(file)
@@ -22,12 +20,10 @@ for file in file_paths:
     output_file_name_temp = output_file_name + ".mp3"
     i = 1
     while os.path.isfile(os.path.basename(output_file_name_temp)):
-        output_file_name_temp = output_file_name + "({}).mp3".format(i)
+        output_file_name_temp = output_file_name + " ({}).mp3".format(i)
         i += 1
     output_file_name = output_file_name_temp
-    try:
-        subprocess.check_call(["ffmpeg", "-i", file, "-codec:a", "libmp3lame", "-qscale:a", "2", output_file_name])
-    except:
+    if subprocess.call(["ffmpeg", "-i", file, "-codec:a", "libmp3lame", "-qscale:a", "2", output_file_name]):
         files_not_converted.append(file_name)
         if os.path.isfile(output_file_name):
             os.remove(output_file_name)
@@ -41,4 +37,4 @@ if files_not_converted:
 else:
     notification_message = "All files were converted successfully."
     notification_icon = "dialog-information"
-notification.Notify('Convert to MP3', 0, notification_icon,"Conversion finished", notification_message,[],{},-1)
+notification.Notify('Convert to MP3', 0, notification_icon, "Conversion finished", notification_message, [], {}, -1)
